Fix page break rules in Apple Books
Currently the fallback rules we put in place for Readium break Apple Books rendering, and removing them obviously breaks Readium. I tried swapping the order but that does nothing.

So this commit adds in tinycss2 to allow us to tag a supports blocks after the qualified rule with the additional Readium fixes, and that doesn’t trigger in Apple Books.